// Del 5/Angular/AdminScreens/src/app/model/Exhibitions/exhibition.ts
export class Exhibition {
    exhibitionID: number;
    exhibitionName: string;
    exhibitionDescription: string;
    exhibitionDate: string;
    exhibitionTime: string;
    exhibition_Image: string;

    exhibitionType: string;

    schedule: string;

    organisation: string;

    exhibitionAnnouncement: string;

    venue: string;
}